requirements.txt
aenum==2.2.3
attrs==19.3.0
beautifulsoup4==4.8.2
bs4==0.0.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
fbchat==1.9.4
gitdb2==2.0.6
GitPython==3.0.5
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.16.0
idna==2.8
numpy==1.18.1
oauthlib==3.1.0
paho-mqtt==1.5.0
pandas==0.25.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyshorteners==0.6.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==3.141.0
six==1.14.0
smmap2==2.0.5
soupsieve==1.9.5
uritemplate==3.0.1
urllib3==1.25.7
